% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmrob2.R
\name{glmrob2.control}
\alias{glmrob2.control}
\title{Auxiliary for Controlling robust GLM Fitting}
\usage{
glmrob2.control(
  acc = 1e-04,
  test.acc = "coef",
  maxit = 50,
  tuning.psi = NULL,
  psi = "bisquare",
  epsilon = 1e-08,
  trace = FALSE
)
}
\arguments{
\item{acc}{see}

\item{test.acc}{see}

\item{maxit}{see}

\item{tuning.psi}{see}

\item{psi}{see}

\item{epsilon}{see}

\item{trace}{see}
}
\description{
Auxiliary for Controlling robust GLM Fitting
}
